+++ Pod/Classes/ViewControllers/ARMapContainerViewController.m
@@ -159,6 +159,7 @@ Since this controller already has to do the above logic, having it handle the Ci
         [self.bottomSheetVC setDrawerPositionWithPosition:[PulleyPosition partiallyRevealed] animated:YES completion:nil];
     } else {
         [self.mapVC setProperty:city.slug forKey:@"citySlug"];
+        [self.mapVC setProperty:@{ @"lat": @(city.epicenter.coordinate.latitude), @"lng": @(city.epicenter.coordinate.longitude) } forKey:@"initialCoordinates"];
         self.initialDataIsLoaded = NO;
     }
 
@@ -213,7 +214,10 @@ Since this controller already has to do the above logic, having it handle the Ci
     CGFloat drawerAbovePartialHeight = [drawer partialRevealDrawerHeightWithBottomSafeArea:bottomSafeArea];
 
     BOOL shouldHideButtons = distance > drawerAbovePartialHeight;
-    [self.mapVC setProperty:@(shouldHideButtons) forKey:@"hideMapButtons"];
+    if (!self.cityPickerController) {
+        // We don't want to unhide buttons if the city picker is on screen.
+        [self.mapVC setProperty:@(shouldHideButtons) forKey:@"hideMapButtons"];
+    }
 }
 
 - (BOOL)fullBleed