+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - Adds zero state to CityTab - kieran
 - Renames SavedShowItemRow to ShowItemRow - kieran
 - Adds Save button functionality to button in All tab - kieran
+- Fixes top buttons re-appearing behind city picker - ash
+- Fixes map not re-centering on new city when changed - ash
 
 ### 1.8.14